Buy/Sell Tooltips; Writing to Windows Registry
Added tooltips that appear when hovering over the Buy or Sell buttons.
These tooltips will tell you what you will have to Buy or Sell at in
order to break even or make a specific % of profit.

The constructor of the main form will now write values to the Windows
registry to try and ensure that the latest version of Internet Explorer
is emulated in the "WebBrowser" control.