Windows Registry Glitch; QoL Change
Fixed a glitch where I was writing a value to the incorrect in the
Windows registry, causing www.bitcoinwisdom.com pages not to load unless
the program was run from within Visual Studio.

Changed the comboxes so that the WEB BROWSER is focused-upon whenever
the exchange changes. This prevents the user from accidently switching
exchanges via the scroll wheel.