Added Capability to Disable Browser
You can now completely disable the built-in browser functionality my
manually resizing the form so-as to hide the browser. This will trigger
internal logic to stop requests and such from being made in the browser.
To further facilitate this functionality, window size and location are
now persistent via the Windows Registry.