Very Basic .ini Support; Custom Tabs
Added very basic support for .ini files. I have essentially only laid
the foundation for a much broader goal which is to ensure the program is
as future-proof as possible.

Started messing around with custom drawing of the tabs encapsulate the
browser.

Made a few minor UI tweaks, restructured a good bit of the code (mainly
BtceApi stuff), and generally promoted stability.